requirements.txt
Authlib
bcrypt
bleach
boto3
click
emoji
Faker
Flask
Flask-Limiter==3.13
flask-login
flask-sock
flask-wtf
gevent
gunicorn
Markdown
minio
peewee==3.18.2
Pillow
psycopg2-binary
pydantic
Pygments
pymdown-extensions
pytest
pytest-cov
pytest-flask
pytest-mock
python-dotenv
python-magic
redis
requests
rich
sentry-sdk[flask]
simple-websocket
